Fixed Minor Observations
