+++ exe/antScript.xml
@@ -1,0 +1,30 @@
+<?xml version="1.0" encoding="UTF-8" standalone="no"?>
+<project default="create_run_jar" name="Create Runnable Jar for Project MessageSender with Jar-in-Jar Loader">
+    <!--this file was created by Eclipse Runnable JAR file Export Wizard-->
+    <!--ANT 1.7 is required-->
+    <!--define folder properties-->
+    <property name="dir.buildfile" value="."/>
+    <property name="dir.workspace" value="${dir.buildfile}/../.."/>
+    <property name="dir.jarfile" value="${dir.buildfile}"/>
+    <target name="create_run_jar">
+        <jar destfile="${dir.jarfile}/MessageSender.jar">
+            <manifest>
+                <attribute name="Main-Class" value="org.eclipse.jdt.internal.jarinjarloader.JarRsrcLoader"/>
+                <attribute name="Rsrc-Main-Class" value="LaunchWindow"/>
+                <attribute name="Class-Path" value="."/>
+                <attribute name="Rsrc-Class-Path" value="./ poi-ooxml-4.0.0.jar xmlbeans-3.0.1.jar poi-ooxml-schemas-4.0.0.jar poi-4.0.0.jar commons-compress-1.18.jar commons-collections4-4.2.jar jackson-annotations-2.9.7.jar jackson-core-2.9.7.jar jackson-databind-2.9.7.jar"/>
+            </manifest>
+            <zipfileset src="jar-in-jar-loader.zip"/>
+            <fileset dir="${dir.workspace}/MessageSender/bin"/>
+            <zipfileset dir="${dir.workspace}/MessageSender/lib" includes="poi-ooxml-4.0.0.jar"/>
+            <zipfileset dir="${dir.workspace}/MessageSender/lib" includes="xmlbeans-3.0.1.jar"/>
+            <zipfileset dir="${dir.workspace}/MessageSender/lib" includes="poi-ooxml-schemas-4.0.0.jar"/>
+            <zipfileset dir="${dir.workspace}/MessageSender/lib" includes="poi-4.0.0.jar"/>
+            <zipfileset dir="${dir.workspace}/MessageSender/lib" includes="commons-compress-1.18.jar"/>
+            <zipfileset dir="${dir.workspace}/MessageSender/lib" includes="commons-collections4-4.2.jar"/>
+            <zipfileset dir="${dir.workspace}/MessageSender/lib" includes="jackson-annotations-2.9.7.jar"/>
+            <zipfileset dir="${dir.workspace}/MessageSender/lib" includes="jackson-core-2.9.7.jar"/>
+            <zipfileset dir="${dir.workspace}/MessageSender/lib" includes="jackson-databind-2.9.7.jar"/>
+        </jar>
+    </target>
+</project>
